Add confirmation to restore-db-dump script

+++ restore-db-dump.sh
@@ -17,28 +17,34 @@ source .env
 DUMP_DIR=$1
 # Gets the newest dump
 FILENAME=$(ls "$DUMP_DIR" | sort | tail -n 1)
+echo -e "\nRestoring from $DUMP_DIR/$FILENAME. Continue? (y/N)"
+read $ANSWER
 
-if [ -n "$2" ]; then
-  DB_CONTAINER=$2
+if [ "$ANSWER" == "y" ]; then
+  if [ -n "$2" ]; then
+    DB_CONTAINER=$2
+  else
+    DB_CONTAINER=cc-mongo
+  fi
+
+  # Copy dump to Docker container
+  sudo docker cp "$DUMP_DIR/$FILENAME" $DB_CONTAINER:/$FILENAME
+  # Extract dump
+  sudo docker exec $DB_CONTAINER sh -c "tar -xvzf /$FILENAME"
+  # Delete archive
+  sudo docker exec $DB_CONTAINER sh -c "rm -f /$FILENAME"
+
+  # Restore collections
+  collections=( "competitions" "people" "rounds" "results" "recordtypes" "events" "schedules" )
+
+  for col in "${collections[@]}"; do
+    sudo docker exec $DB_CONTAINER mongorestore -u $MONGO_DEV_USERNAME -p $MONGO_DEV_PASSWORD --db cubingcontests -c $col /dump/cubingcontests/$col.bson
+  done
+
+  # Delete dump directory
+  sudo docker exec $DB_CONTAINER sh -c "rm -rf /dump"
+
+  echo -e "\nDone!"
 else
-  DB_CONTAINER=cc-mongo
+  echo -e "\nAborting"
 fi
-
-# Copy dump to Docker container
-sudo docker cp "$DUMP_DIR/$FILENAME" $DB_CONTAINER:/$FILENAME
-# Extract dump
-sudo docker exec $DB_CONTAINER sh -c "tar -xvzf /$FILENAME"
-# Delete archive
-sudo docker exec $DB_CONTAINER sh -c "rm -f /$FILENAME"
-
-# Restore collections
-collections=( "competitions" "people" "rounds" "results" "recordtypes" "events" "schedules" )
-
-for col in "${collections[@]}"; do
-  sudo docker exec $DB_CONTAINER mongorestore -u $MONGO_DEV_USERNAME -p $MONGO_DEV_PASSWORD --db cubingcontests -c $col /dump/cubingcontests/$col.bson
-done
-
-# Delete dump directory
-sudo docker exec $DB_CONTAINER sh -c "rm -rf /dump"
-
-echo -e "\nDone!"
